requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langsmith
fastapi
uvicorn
python-dotenv
langchain-chroma
langchain-huggingface
sentence-transformers
pydantic>=2.0
pydantic-settings
requests
langchain-community
langchain-openai
rank_bm25
watchdog
einops
huggingface_hub
